requirements.txt
aiohttp==3.8.4
fastapi==0.110.0
fastapi-utils==0.2.1
pydantic==1.10.12
pytz==2023.3
requests==2.31.0
requests_oauthlib==1.3.1
git+https://JENKINSKEY@github.alm.europe.cloudcenter.corp/sanes-shuttle/shuttlelib.git@master
six==1.16.0
urllib3==1.26.10
uvicorn==0.18.2
influxdb_client==1.41.0
kubernetes==27.2.0
websocket-client==1.6.1
email-validator==2.0.0.post2
openshift-client==1.0.19